Linted and fixed several files.

+++ src/include/einsums/_Common.hpp
@@ -29,44 +29,52 @@
 
 #pragma once
 
+#include <array>
+#include <cstdint>
+#include <ostream>
+#include <string>
+#include <utility>
+
 #include "einsums/Error.hpp"
 #include "einsums/Print.hpp"
 #include "einsums/_Compiler.hpp"
 #include "einsums/_Export.hpp"
 
-#include <array>
-#include <cstdint>
-#include <ostream>
-
+// Macro definitions
 #define EINSUMS_STRINGIFY(a)  EINSUMS_STRINGIFY2(a)
 #define EINSUMS_STRINGIFY2(a) #a
 
-#define BEGIN_EINSUMS_NAMESPACE_HPP(x)                                                                                                     \
-    namespace x {                                                                                                                          \
-    namespace detail {                                                                                                                     \
-    extern EINSUMS_EXPORT std::string s_Namespace;                                                                                         \
+// Begin the namespace and include details.
+#define BEGIN_EINSUMS_NAMESPACE_HPP(x)                                         \
+    namespace x {                                                              \
+    namespace detail {                                                         \
+    extern EINSUMS_EXPORT std::string s_Namespace;                             \
     }
 
+// End the namespace.
 #define END_EINSUMS_NAMESPACE_HPP(x) }
 
-#define BEGIN_EINSUMS_NAMESPACE_CPP(x)                                                                                                     \
-    namespace x {                                                                                                                          \
-    namespace detail {                                                                                                                     \
-    EINSUMS_EXPORT std::string s_Namespace = #x;                                                                                           \
+// Begin the namespace and include details.
+#define BEGIN_EINSUMS_NAMESPACE_CPP(x)                                         \
+    namespace x {                                                              \
+    namespace detail {                                                         \
+    EINSUMS_EXPORT std::string s_Namespace = #x;                               \
     }
 
+// End the namespace.
 #define END_EINSUMS_NAMESPACE_CPP(x) }
 
 namespace einsums {
 
+// Define types for later use.
 #if defined(MKL_ILP64)
-using eint  = long long int;
-using euint = unsigned long long int;
-using elong = long long int;
+using eint  = int128_t;   // long long int;
+using euint = uint128_t;  // unsigned long long int;
+using elong = int128_t;   // long long int;
 #else
 using eint  = int;
 using euint = unsigned int;
-using elong = long int;
+using elong = int64_t;    // long int;
 #endif
 
 // Functions to start and end processing of tensor operations.
@@ -89,7 +97,9 @@ auto EINSUMS_EXPORT initialize() -> int;
  */
 void EINSUMS_EXPORT finalize(bool timerReport = false);
 
-// The following detail and "using" statements below are needed to ensure Dims, Strides, and Offsets are strong-types in C++
+/* The following detail and "using" statements below are needed to ensure Dims,
+ * Strides, and Offsets are strong-types in C++
+ */
 namespace detail {
 
 struct DimType {};
@@ -114,10 +124,12 @@ struct Array : public std::array<UnderlyingType, Rank> {
      * @callergraph
      */
     template <typename... Args>
-    constexpr explicit Array(Args... args) : std::array<UnderlyingType, Rank>{static_cast<UnderlyingType>(args)...} {}
+    constexpr explicit Array(Args... args) :
+            std::array<UnderlyingType, Rank>{
+                static_cast<UnderlyingType>(args)...} {}
     using type = T;
 };
-} // namespace detail
+}  // namespace detail
 
 /**
  * @class Dim
@@ -172,7 +184,7 @@ struct All_t {};
  */
 static struct All_t All;
 
-} // namespace einsums
+}  // namespace einsums
 
 /**
  * Prints the dimensions passed in.
